Dropzone added

+++ app/Http/Controllers/CommentsController.php
@@ -37,7 +37,10 @@ class CommentsController extends Controller
 
     public function update(Comment $comment)
     {
-        // TODO: Only owner can update
+        if(!$comment->author())
+        {
+            return abort(403);
+        }
 
         $comment->update([
             'body' => request('body')
@@ -48,7 +51,10 @@ class CommentsController extends Controller
 
     public function destroy(Comment $comment)
     {
-        // TODO: Only owner can delete
+        if(!$comment->author())
+        {
+            return abort(403);
+        }
 
         $comment->delete();
